+++ src/ft_string/ft_strspn.c
@@ -6,15 +6,15 @@
 /*   By: ecarvalh <ecarvalh@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/12 18:46:45 by ecarvalh          #+#    #+#             */
-/*   Updated: 2024/02/21 12:35:11 by ecarvalh         ###   ########.fr       */
+/*   Updated: 2024/02/26 16:13:59 by ecarvalh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strspn(char const *str, char const *accept)
+size_t	ft_strspn(char const *str, char const *accept)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] && ft_strchr(accept, str[i]))
